add helpers for ksp types
This CL adds helpers to convert a KSTypeReference to a typeName.

Also changed the KspProcesisngEnv to also receive the resolver.

For convenient local debugging, i added an environment variable to
only run ksp tests. It is temporary until KSP is merged so not making
it generic. Similarly, added APIs to run tests just with KSP. We get
a lot from Auto Common where we need to duplicate for KSP so it makes
sense to have KSP only test infra here.

Test: KSTypeExtTest