add helpers for ksp types (#2)
* add helpers for ksp types

This CL adds helpers to convert a KSTypeReference to a typeName.

Also changed the KspProcesisngEnv to also receive the resolver.

For convenient local debugging, i added an environment variable to
only run ksp tests. It is temporary until KSP is merged so not making
it generic. Similarly, added APIs to run tests just with KSP. We get
a lot from Auto Common where we need to duplicate for KSP so it makes
sense to have KSP only test infra here.

Test: KSTypeExtTest

* minor code cleanup

* ktlint fixes

* make KSType.typeName nullable

For cases where we fallback to the reference, it is better to handle
it in the outer loop as there are cases where we might end up w/ a
KSType so we should be able to call typeName without a reference

* address review comments

* add notes for the KSName issues w/ packages

Also changed the generation to assume KSName models simple names
separately

* use new package api